Add non regression tests based on homologation test cases
The idea is to parse homologation command and to generate the SOAP xml
corresponding to each command and test against a pre-generated xml.

So in case we change something in the code we get notified by a
potentially failing test.

To regenerate test data we can run "pytest --regen-test-data"

I had to hack a bit some date/default which are evaluated at import time
to make this work.